+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The major version bump is due to dropped support for PHP `8.1` and a several bre
 * The development error view now displays the Mako environment name.
 * Added `Connection::blob()` method that allows you to easily fetch a blob column as stream.
 * Added `Query::blob()` method that allows you to easily fetch a blob column as stream.
+* The `ManyToMany::unlink()` and ManyToMany::synchronize()` methods now support junction attributes.
 
 #### Improvements